requirements.txt
git+https://github.com/qubvel/segmentation_models.pytorch
Flask==2.3.2
Pillow
opencv-python-headless==4.5.5.64
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.1.3
matplotlib==3.6.3
seaborn==0.12.1
tqdm==4.64.1
scikit-image==0.19.3
